requirements.txt
moviepy==1.0.3
requests
python-dotenv
pydub
edge-tts
Pillow
openai-whisper
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
